+++ sqlstore/src/main/resources/db/migration/V8505__transfer_request_name.sql
@@ -1,0 +1,1 @@
+ALTER TABLE Transfer ADD COLUMN transferRequestName varchar(100);